// backend/src/reservations/enums/reservation.enum.ts
export enum ReservationPurpose {
  COURSE = 'course',
  REHEARSAL = 'rehearsal',
  COMPETITION = 'competition',
  CLUB = 'club',
  OTHER = 'other',
}

export enum ReservationStatus {
  PENDING = 'pending',
  APPROVED = 'approved',
  REJECTED = 'rejected',
  CANCELLED = 'cancelled',
  BORROWED = 'borrowed',
  RETURNED = 'returned',
  COMPLETED = 'completed',
}